Add capacity for anonymous ordering, add credit card tracking to orders and users

+++ app/controllers/orders_controller.rb
@@ -1,7 +1,7 @@
 class OrdersController < ApplicationController
   before_filter :skip_mini_cart
   before_filter :find_order, only: [:show, :destroy]
-  before_filter :require_login, only: [:create, :index, :show]
+  before_filter :require_login, only: [:index, :show]
 
   def show
 
@@ -16,13 +16,14 @@ class OrdersController < ApplicationController
   end
 
   def create
-    @order = current_user.orders.create!(status: 'pending')
-    @order.attach_addresses(params)
+    @order = Order.create(params.merge(status: 'pending'))
+    @order.user_id = current_user.id if current_user
+    return redirect_to :back unless @order.save
     current_cart.cart_items.each do |cart_item|
       @order.order_items.create(product_id: cart_item.product_id,
         quantity: cart_item.quantity)
     end
-    current_user.update_addresses(params)
+    current_user.update_addresses(params) if current_user
     current_cart.destroy
   end
